+++ experiments/billable/scripts/lib/vitePlugins/miniflarePlugin/plugin.mts
@@ -3,7 +3,7 @@ import { readFile } from "node:fs/promises";
 import { EventEmitter } from "node:events";
 import { fileURLToPath } from "node:url";
 import { resolve } from "node:path";
-import { createRequire } from "node:module";
+import { createRequire, builtinModules } from "node:module";
 
 import { resolve as importMetaResolve } from "import-meta-resolve";
 import colors from "picocolors";
@@ -34,6 +34,17 @@ import {
 import { compileTsModule } from "../../compileTsModule.mjs";
 import { getShortName } from "../../getShortName.mjs";
 import { SRC_DIR } from "../../constants.mjs";
+
+export const CLOUDFLARE_BUILT_IN_MODULES = [
+  'cloudflare:email',
+  'cloudflare:sockets',
+  'cloudflare:workers',
+];
+
+export const NODE_BUILT_IN_MODULES = [
+  ...builtinModules,
+  ...builtinModules.map(module => `node:${module}`)
+]
 
 interface MiniflarePluginOptions {
   entry: string;
@@ -391,6 +402,7 @@ export const miniflarePlugin = async (
               platform: "browser",
               banner: undefined,
             },
+            exclude: [...CLOUDFLARE_BUILT_IN_MODULES, ...NODE_BUILT_IN_MODULES],
           },
           build: {
             ssr: true,
@@ -398,6 +410,7 @@ export const miniflarePlugin = async (
               input: {
                 index: entry,
               },
+              external: [...CLOUDFLARE_BUILT_IN_MODULES, ...NODE_BUILT_IN_MODULES],
             },
           },
         },
